+++ src/local_flight_map/api/base/config.py
@@ -1,0 +1,17 @@
+
+from pydantic_settings import BaseSettings, SettingsConfigDict
+
+
+class BaseConfig(BaseSettings):
+    """
+    Base configuration class for API clients.
+    Uses pydantic for configuration management.
+    """
+    model_config = SettingsConfigDict(
+        env_file=".env",
+        env_file_encoding="utf-8",
+        env_nested_delimiter="_",
+        frozen=True,
+        extra="ignore",
+        case_sensitive=False
+    )
